update modal and uploading
